Generified some behaviour and added daily rewards multiplier
